Update all suite in DatabaseApi/Assets, add asset object validation, add GetAssetObjects suite
