+++ common/object_validation.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 import lemoncheesecake.api as lcc
-from lemoncheesecake.matching import has_length, is_integer, check_that_in, check_that, is_dict, is_list
+from lemoncheesecake.matching import has_length, is_integer, check_that_in, check_that, is_dict, is_list,\
+    require_that, is_str
 
 
 class ObjectValidator(object):
@@ -74,3 +75,73 @@ class ObjectValidator(object):
                     "extensions", is_list(),
                     quiet=True
                 )
+
+    @staticmethod
+    def validate_asset_structure(base_test, asset):
+
+        def validate_core_exchange_rate_structure(base_test, core_exchange_rate):
+            check_that_in(
+                core_exchange_rate,
+                "base", is_dict(),
+                "quote", is_dict(),
+                quiet=True
+            )
+            for key in core_exchange_rate:
+                core_exchange_rate_part = core_exchange_rate[key]
+                base_test.check_uint64_numbers(core_exchange_rate_part, "amount", quiet=True)
+                if not base_test.type_validator.is_asset_id(core_exchange_rate_part["asset_id"]):
+                    lcc.log_error("Wrong format of {} 'asset_id', got: {}".format(
+                        key, core_exchange_rate_part["asset_id"]))
+                else:
+                    lcc.log_info("{} 'asset_id' has correct format: asset_id".format(key))
+
+        require_that(
+            "'length of chain asset'",
+            asset, has_length(7)
+        )
+
+        if not base_test.type_validator.is_asset_id(asset["id"]):
+            lcc.log_error("Wrong format of 'id', got: {}".format(asset["id"]))
+        else:
+            lcc.log_info("'id' has correct format: asset_id")
+        if not base_test.type_validator.is_dynamic_asset_data_id(asset["dynamic_asset_data_id"]):
+            lcc.log_error("Wrong format of 'dynamic_asset_data_id', got: {}".format(
+                asset["dynamic_asset_data_id"]))
+        else:
+            lcc.log_info("'dynamic_asset_data_id' has correct format: dynamic_asset_data_id")
+
+        if not base_test.type_validator.is_account_id(asset["issuer"]):
+            lcc.log_error("Wrong format of 'issuer', got: {}".format(asset["issuer"]))
+        else:
+            lcc.log_info("'issuer' has correct format: account_id")
+        if not base_test.type_validator.is_asset_name(asset["symbol"]):
+            lcc.log_error("Wrong format of 'symbol', got: {}".format(asset["symbol"]))
+        else:
+            lcc.log_info("'symbol' has correct format: asset_name")
+        check_that_in(
+            asset,
+            "options", is_dict(),
+            "extensions", is_list(),
+            "precision", is_integer(8),
+            quiet=True
+        )
+        options = asset["options"]
+        require_that("'options'", options, has_length(8))
+        check_that_in(
+            options,
+            "blacklist_authorities", is_list(),
+            "core_exchange_rate", is_dict(),
+            "description", is_str(),
+            "extensions", is_list(),
+            "flags", is_integer(),
+            "issuer_permissions", is_integer(),
+            "whitelist_authorities", is_list(),
+            quiet=True
+        )
+        core_exchange_rate = options["core_exchange_rate"]
+        require_that(
+            "'core_exchange_rate'",
+            core_exchange_rate, has_length(2)
+        )
+        validate_core_exchange_rate_structure(base_test, core_exchange_rate)
+        base_test.check_uint64_numbers(options, "max_supply", quiet=True)
